[KERNELBASE] Build parts of kernelbase [KERNELBASE] Get loader.c building [KERNELBASE] Volume.c now builds [WINE] Add some definitions to winternl. [PSDK] Add some missing constants to winnls.h [KERNEL32] Add stubs for HeapAlloc and HeapFree
This allows external code linked into kernel32 to link to these functions, which are otherwise defined as macros to RtlAllocateHeap and RtlFreeHeap.
This will be needed for NT6 apis.
[KERNELBASE] build locale.
[KERNELBASE] build path.c
[WIN32_REGISTRY] Add win32_registry static library

To be used by NT6 apis in kernel32

This is not great. It works, but it puts stuff into the lib that shouldn't be in there. We need it, because the includes need the generated files.
Maybe a dependency could solve that.

[KERNELBASE] build thread.c